// src/common/enums/leave-type.enum.ts
export enum LeaveType {
  COMPENSATORY = 'compensatory', // Nghỉ bù
  PAID  = 'paid', // Nghỉ phép (có lương)
  UNPAID = 'unpaid', // Nghỉ không lương
  PERSONAL_PAID = 'personal_paid', // Nghỉ việc riêng (vẫn được trả lương)
  INSURANCE = 'insurance', // Nghỉ bảo hiểm (ốm đau, thai ...)
}